+++ Framework-Headers/PrivateFrameworks/SpotlightServices._framework/Objc/SSCoreMLInterface.h
@@ -19,7 +19,7 @@
 - (id)getVersionString;
 - (id)initSpecificModel:(id)arg1 error:(id *)arg2;
 - (id)initWithURL:(id)arg1 error:(id *)arg2;
-- (double)predict:(id)arg1 error:(id *)arg2 isL2:(BOOL)arg3;
+- (float)predict:(id)arg1 error:(id *)arg2 isL2:(BOOL)arg3;
 
 @end
